Nuking unused codez (#120)
* Nuking

* Nuking of old realm bridge nonsense

* More deletions, fixing da build

* More deleting of things

* Goodbye old modules my old friend

+++ Canvas/CanvasKit1/API Classes/CKTerm.m
@@ -47,7 +47,7 @@
 }
 
 - (NSUInteger)hash {
-    return self.ident;
+    return (NSUInteger)self.ident;
 }
 
 @end
