+++ Canvas/CanvasKit1/Models/CKAssignmentGroup.m
@@ -79,7 +79,7 @@
 }
 
 - (NSUInteger)hash {
-    return ident;
+    return (NSUInteger)ident;
 }
 
 @end
