+++ Canvas/CanvasKit1/Models/CKAssignmentOverride.m
@@ -103,7 +103,7 @@
 
 
 - (NSUInteger)hash {
-    return _ident;
+    return (NSUInteger)_ident;
 }
 
 - (NSString *)description {
