+++ Canvas/CanvasKit1/Models/CKConversationRelatedSubmission.m
@@ -58,7 +58,7 @@
 }
 
 - (NSUInteger)hash {
-    return assignmentIdent ^ userIdent;
+    return (NSUInteger)assignmentIdent ^ userIdent;
 }
 
 @end
