+++ Canvas/CanvasKit1/Models/CKDiscussionTopic.m
@@ -136,7 +136,7 @@
 }
 
 - (NSUInteger)hash {
-    return ident;
+    return (NSUInteger)ident;
 }
 
 @end
