+++ Canvas/CanvasKit1/Models/CKEnrollment.m
@@ -146,7 +146,7 @@
 }
 
 - (NSUInteger)hash {
-    return ident;
+    return (NSUInteger)ident;
 }
 
 @end
