+++ Canvas/CanvasKit1/Models/CKGroup.m
@@ -141,7 +141,7 @@ static CKGroupRole groupRoleFromString(NSString *string) {
 }
 
 - (NSUInteger)hash {
-    return self.ident;
+    return (NSUInteger)self.ident;
 }
 
 @end
