+++ Canvas/CanvasKit1/Models/CKStudent.m
@@ -71,7 +71,7 @@
 }
 
 - (NSUInteger)hash {
-    return self.ident;
+    return (NSUInteger)self.ident;
 }
 
 @end
