+++ Canvas/CanvasKit1/Models/CKSubmission.m
@@ -340,7 +340,7 @@
 }
 
 - (NSUInteger)hash {
-    return self.ident;
+    return (NSUInteger)self.ident;
 }
 
 @end
