+++ Canvas/TechDebt/MKNumberBadgeView/MKNumberBadgeView.m
@@ -122,7 +122,7 @@
 	
 	CGContextRef curContext = UIGraphicsGetCurrentContext();
 
-	NSString* numberString = [NSString stringWithFormat:@"%d",self.value];
+	NSString* numberString = [NSString stringWithFormat:@"%lu",(unsigned long)self.value];
 	
 	
 	CGSize numberSize = [numberString sizeWithFont:self.font];
@@ -281,7 +281,7 @@
 
 - (CGSize)badgeSize
 {
-	NSString* numberString = [NSString stringWithFormat:@"%d",self.value];
+	NSString* numberString = [NSString stringWithFormat:@"%lu",(unsigned long)self.value];
 	
 	
 	CGSize numberSize = [numberString sizeWithFont:self.font];
